Redefine the relative monad on the J^M functor.
Also redefine the J^M functor in line with the rework of the general
theory.
Compared to the previous definition this one doesn't bake in the
notion of extension of contexts and doesn't have extensions of
metacontexts at all. The definiion is now much simpler and should
allow us to express the relaton between the two relative monads (the
context and metacontext ones).